requirements.txt
Appium-Python-Client==1.2.0
attrs==21.2.0
flake8==3.9.2
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
selenium==3.141.0
toml==0.10.2
urllib3==1.26.6
